requirements.txt
# ============================================================
# Gaussian Color Adjust — dipendenze Python
# ============================================================
#
# INSTALLAZIONE STANDARD (macOS Apple Silicon MPS):
#   pip install -r requirements.txt
#   Nota: gsplat viene installato ma NON funzionerà su MPS (kernel CUDA assenti).
#   Il progetto usa automaticamente il renderer PyTorch puro come fallback.
#
# INSTALLAZIONE ROCm (Linux, AMD RX 9070 XT):
#   1. Installa prima PyTorch ROCm (vedi requirements_rocm.txt)
#   2. pip install -r requirements.txt
#   3. Per gsplat con ROCm: vedi requirements_rocm.txt
#
# INSTALLAZIONE NVIDIA CUDA:
#   pip install -r requirements.txt   # gsplat si installa con CUDA automaticamente
# ============================================================

# Core PyTorch
torch>=2.1.0
torchvision>=0.16.0

# Dati e immagini
numpy>=1.24.0
Pillow>=10.0.0
plyfile>=0.9
tqdm>=4.65.0

# Gaussian Splatting rasterizer (richiede CUDA o ROCm compilato)
# Su Apple MPS / CPU: viene usato automaticamente il renderer PyTorch puro
# Su NVIDIA/AMD: pip install gsplat  oppure vedi requirements_rocm.txt per AMD
gsplat>=1.0.0

# Diffusion models
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0

# Integrazione COLMAP (opzionale)
pycolmap>=0.5.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
